Move functions file. #206
